move all Strucutre functions inside FinancialStructureManager and delete Structure
